Rearranging things for better structure
