+++ src/app/components/Composer/index.js
@@ -1,12 +1,21 @@
 // tools
 import React from "react"
 import PropTypes from "prop-types"
-import { Editor } from "slate"
-import { composerInputPlaceholders, plugins, html, stateSchema } from "./constants"
+import { Editor, Html } from "slate"
+
+import { plugins } from "./plugins"
+import { schema } from "./schema"
 
 // styles
 import { PageHeader, PageByline } from "../Article/styles/header"
 import { PageTitle, PageSubtitle } from "./styles"
+
+// state
+import initialState from "./state.json"
+
+// constants
+const rules = schema.rules
+const html = new Html({ rules })
 
 
 
@@ -14,17 +23,16 @@ import { PageTitle, PageSubtitle } from "./styles"
 export const ComposerHead = props => {
 	return (
 		<PageHeader>
-			<PageTitle placeholder={ 			composerInputPlaceholders.title }></PageTitle>
-			<PageSubtitle placeholder={ 	composerInputPlaceholders.subtitle }></PageSubtitle>
+			<PageTitle placeholder={ 			initialState.title }></PageTitle>
+			<PageSubtitle placeholder={ 	initialState.subtitle }></PageSubtitle>
 			<PageByline>Your name & profile link will appear here.</PageByline>
 		</PageHeader>
 	)
 }
-
 export class ComposerBody extends React.Component {
 	state = {
-		state: html.deserialize(localStorage.getItem("composer-state") || composerInputPlaceholders.body),
-		schema: stateSchema
+		state: html.deserialize(localStorage.getItem("composer-state") || initialState.body),
+		schema
 	}
   onChange = state => this.setState({ state })
   onPaste = (e, data, state) => {
