+++ src/app/components/Composer/plugins/index.js
@@ -1,7 +1,21 @@
-// tools
-
 // components
-import { PluginMarkHotkey } from "./mark-hotkey"
+import { MarkHotkey } from "./mark-hotkey"
+import AutoReplace from "slate-auto-replace"
+import EditBlockquote from "slate-edit-blockquote"
 
-// return
-export const MarkHotkey = PluginMarkHotkey
+// export
+export const plugins = [
+  MarkHotkey({ 	key: "b", 			type: "bold" }),
+  MarkHotkey({ 	key: "i", 			type: "italic" }),
+	EditBlockquote({
+		type: 				"quote",
+		typeDefault: 	"paragraph",
+	}),
+  AutoReplace({
+    trigger: 		"space",
+    before: 		/^(>)$/,
+    transform: 	(transform, e, data, matches) => {
+      return transform.setBlock({ type: 'quote' })
+    }
+  })
+]
